refactor: Move some classes into contexts & errors folders.
